feat(tasks): extract filter toolbar
