+++ dashboard-ui/app/tasks/TasksContent.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import { http, HttpResponse } from 'msw'
 import { vi } from 'vitest'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import TasksTable from './TasksTable'
+import TasksContent from './TasksContent'
 import { server } from '@/tests/mocks/server'
 
 const renderWithClient = (ui: React.ReactElement) => {
@@ -13,7 +13,7 @@ const renderWithClient = (ui: React.ReactElement) => {
   )
 }
 
-describe('TasksTable', () => {
+describe('TasksContent', () => {
   beforeAll(() => {
     vi.setSystemTime(new Date('2024-01-05'))
   })
@@ -23,7 +23,7 @@ describe('TasksTable', () => {
   })
 
   it('renders tasks from API (happy path)', async () => {
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     expect(await screen.findByText('Task 1')).toBeInTheDocument()
   })
 
@@ -31,7 +31,7 @@ describe('TasksTable', () => {
     server.use(
       http.get('http://localhost:4000/api/task', () => HttpResponse.error())
     )
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     await waitFor(() => expect(screen.getByText(/error/i)).toBeInTheDocument())
   })
 
@@ -39,7 +39,7 @@ describe('TasksTable', () => {
     server.use(
       http.get('http://localhost:4000/api/task', () => HttpResponse.json([]))
     )
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     await waitFor(() => {
       const rows = screen.queryAllByRole('row')
       // header + no data
@@ -48,7 +48,7 @@ describe('TasksTable', () => {
   })
 
   it('filters by priority', async () => {
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     await screen.findByText('Task 1')
     const select = screen.getByRole('combobox', { name: /приоритет/i })
     await userEvent.selectOptions(select, 'Высокий')
@@ -56,7 +56,7 @@ describe('TasksTable', () => {
   })
 
   it('deletes a task', async () => {
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     const menuBtn = await screen.findByRole('button', { name: /действия/i })
     await userEvent.click(menuBtn)
     const deleteItem = await screen.findByRole('menuitem', { name: /удалить/i })
@@ -70,7 +70,7 @@ describe('TasksTable', () => {
   })
 
   it('opens task info modal on row click and closes with ESC', async () => {
-    renderWithClient(<TasksTable />)
+    renderWithClient(<TasksContent />)
     const cell = await screen.findByText('Task 1')
     await userEvent.click(cell)
     const dialog = await screen.findByRole('dialog', { name: /task 1/i })
